+++ packages/rsc/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @ai-sdk/rsc
 
+## 3.0.0-beta.84
+
+### Patch Changes
+
+- ai@7.0.0-beta.83
+
+## 3.0.0-beta.83
+
+### Patch Changes
+
+- Updated dependencies [e27ed76]
+  - ai@7.0.0-beta.82
+
+## 3.0.0-beta.82
+
+### Patch Changes
+
+- Updated dependencies [2fe1099]
+- Updated dependencies [f04adcb]
+  - ai@7.0.0-beta.81
+
+## 3.0.0-beta.81
+
+### Patch Changes
+
+- Updated dependencies [3ae1786]
+  - @ai-sdk/provider-utils@5.0.0-beta.17
+  - ai@7.0.0-beta.80
+
+## 3.0.0-beta.80
+
+### Patch Changes
+
+- Updated dependencies [6866afe]
+  - ai@7.0.0-beta.79
+
+## 3.0.0-beta.79
+
+### Patch Changes
+
+- Updated dependencies [f372547]
+  - ai@7.0.0-beta.78
+
 ## 3.0.0-beta.78
 
 ### Patch Changes